doing naive feature extraction firsh

+++ models/slic_vit.py
@@ -381,7 +381,7 @@ class SLICViT(nn.Module):
         #text_features_normalized = (text_features - torch.min(text_features)) / (torch.max(text_features) - torch.min(text_features))
         #text_features_normalized = torch.tensor(np.load(root_path + "Nesf0_2D/" + text + "_clip_feature.npy"))
         #print(text_features_normalized)
-        # text_features_normalized = text_features_normalized.to(torch.float)
+        text_features_normalized = text_features_normalized.to(torch.float)
         # image_features_normalized = image_features_normalized.to(torch.float)
         for i in range(r):
             for j in range(c):
